backend/requirements.txt
fastapi
uvicorn
websockets
httpx
python-dotenvProcfile
jinja2
python-multipart